#51 Move StateFactory to the state package
